MessageReceiver.hs, hlint, relo single-use funs
Move message receiver into its own modules with all necc. functions;
Move single-use functions out of Types.hs;
hlint cleanups: unnecc. parens, unnecc. bang patterns, use guards
