+++ src/Juno/Util/Util.hs
@@ -5,14 +5,13 @@ module Juno.Util.Util
   ( seqIndex
   , lastLogInfo
   , getQuorumSize
-  , debug
+  , debug, debugNoRole
   , randomRIO
   , runRWS_
   , enqueueEvent, enqueueEventLater
   , dequeueEvent
   , logMetric
   , logStaticMetrics
-  , messageReceiver
   , setTerm
   , setRole
   , setCurrentLeader
@@ -25,10 +24,7 @@ module Juno.Util.Util
 import Juno.Runtime.Types
 import Juno.Util.Combinator
 
-import Data.Either (partitionEithers)
-import Data.List (partition)
 import Control.Lens
-import Control.Parallel.Strategies
 import Data.Sequence (Seq)
 import Control.Monad.RWS
 import Data.ByteString (ByteString)
@@ -37,7 +33,6 @@ import qualified Control.Concurrent.Lifted as CL
 import qualified Data.ByteString as B
 import qualified Data.Sequence as Seq
 import qualified Data.Map.Strict as Map
-import qualified Data.Set as Set
 import qualified System.Random as R
 
 seqIndex :: Seq a -> Int -> Maybe a
@@ -77,7 +72,7 @@ randomRIO :: (Monad m, R.Random a) => (a,a) -> Raft m a
 randomRIO rng = view (rs.random) >>= \f -> f rng -- R.randomRIO
 
 runRWS_ :: Monad m => RWST r w s m a -> r -> s -> m ()
-runRWS_ ma r s = runRWST ma r s >> return ()
+runRWS_ ma r s = void $ runRWST ma r s
 
 -- no state update
 enqueueEvent :: Monad m => Event -> Raft m ()
@@ -100,62 +95,6 @@ logStaticMetrics = do
   logMetric . MetricClusterSize =<< view clusterSize
   logMetric . MetricQuorumSize =<< view quorumSize
 
--- | Thread to take incoming messages and write them to the event queue.
--- THREAD: MESSAGE RECEIVER (client and server), no state updates
-messageReceiver :: Monad m => Raft m ()
-messageReceiver = do
-  gm <- view (rs.getMessages)
-  getCmds <- view (rs.getNewCommands)
-  getAers <- view (rs.getNewEvidence)
-  ks <- KeySet <$> view (cfg.publicKeys) <*> view (cfg.clientPublicKeys)
-  forever $ do
-    -- NB: This all happens on one thread because it runs in Raft and we're trying (too hard) to avoid running in IO
-
-    -- Take a big gulp of AERs, the more we get the more we can skip
-    (howManyAers, alotOfAers, invalidAers) <- toAlotOfAers <$> getAers 2000
-    unless (alotOfAers == mempty) $ do debugNoRole $ "Combined together " ++ show howManyAers ++ " AERs"
-                                       enqueueEvent $ AERs alotOfAers
-    mapM_ debugNoRole invalidAers
-    -- sip from the general message stream, this should be relatively underpopulated except during an election but can contain HUGE AEs
-    gm 50 >>= sequentialVerify ks
-    -- now take a massive gulp of commands
-    verifiedCmds <- parallelVerify ks <$> getCmds 5000
-    (invalidCmds, validCmds) <- return $ partitionEithers verifiedCmds
-    mapM_ debugNoRole invalidCmds
-    cmds@(CommandBatch cmds' _) <- return $ batchCommands validCmds
-    lenCmdBatch <- return $ length cmds'
-    unless (lenCmdBatch == 0) $ do
-      enqueueEvent $ ERPC $ CMDB' cmds
-      debugNoRole $ "AutoBatched " ++ show (length cmds') ++ " Commands"
-
-toAlotOfAers :: [(ReceivedAt,SignedRPC)] -> (Int, AlotOfAERs, [String])
-toAlotOfAers s = (length decodedAers, alotOfAers, invalids)
-  where
-    (invalids, decodedAers) = partitionEithers $ uncurry aerOnlyDecode <$> s
-    mkAlot aer@AppendEntriesResponse{..} = AlotOfAERs $ Map.insert _aerNodeId (Set.singleton aer) Map.empty
-    alotOfAers = mconcat (mkAlot <$> decodedAers)
-
-sequentialVerify :: Monad m => KeySet -> [(ReceivedAt, SignedRPC)] -> Raft m ()
-sequentialVerify ks msgs = do
-  (aes, noAes) <- return $ partition (\(_,SignedRPC{..}) -> if _digType _sigDigest == AE then True else False) msgs
-  (invalid, validNoAes) <- return $ partitionEithers $ parallelVerify ks noAes
-  mapM_ (enqueueEvent . ERPC) validNoAes
-  mapM_ debugNoRole invalid
-  -- AE's have the potential to be BIG so we need to take care not to do them in parallel by accident
-  mapM_ (\(ts,msg) -> case signedRPCtoRPC (Just ts) ks msg of
-            Left err -> debugNoRole err
-            Right v -> enqueueEvent $ ERPC v) aes
-
-parallelVerify :: KeySet -> [(ReceivedAt, SignedRPC)] -> [Either String RPC]
-parallelVerify ks msgs = ((\(ts, msg) -> signedRPCtoRPC (Just ts) ks msg) <$> msgs) `using` parList rseq
-
-batchCommands :: [RPC] -> CommandBatch
-batchCommands cmdRPCs = cmdBatch
-  where
-    cmdBatch = CommandBatch (concat (prepCmds <$> cmdRPCs)) NewMsg
-    prepCmds (CMD' cmd) = [cmd]
-    prepCmds (CMDB' (CommandBatch cmds _)) = cmds
-    prepCmds o = error $ "Invariant failure in batchCommands: " ++ show o
 
 setTerm :: Monad m => Term -> Raft m ()
 setTerm t = do
